* Fix compilation errors due to interface changes in Xpert.ivy. (64 bit database identifiers) - Reviewed by nobody
Former-commit-id: fa7a48093c67cbe983d042d84a4e4fb45a57d472

+++ ivyProjects/IvyAddOns/src/ch/ivyteam/ivy/addons/restricted/eventlog/EventLogHelper.java
@@ -348,7 +348,7 @@ public class EventLogHelper
    * @return the resulting list
    * @throws PersistencyException
    */
-  public static synchronized List<EventLog> readAllByCase(Integer caseId, IWorkflowContext wf)
+  public static synchronized List<EventLog> readAllByCase(Long caseId, IWorkflowContext wf)
           throws PersistencyException
   {
     List<EventLog> result;
